+++ tests/Fixtures/resources/laragram/reply_view/reply_keyboard.php
@@ -1,5 +1,3 @@
-<?php
-
 resize();
 
 reply('Option A');